+++ lib/ansible/modules/cloud/alicloud/ali_instance.py
@@ -694,7 +694,9 @@ def main():
     if instance_ids:
         if not isinstance(instance_ids, list):
             module.fail_json(msg='The parameter instance_ids should be a list, aborting')
-        instances = ecs.describe_instances(zone_id=zone_id, instance_ids=instance_ids)
+        total_ids = len(instance_ids)
+        pagesize = total_ids if 10 < total_ids <= 100 else (100 if total_ids > 100 else 10)
+        instances = ecs.describe_instances(zone_id=zone_id, instance_ids=instance_ids, pagesize=pagesize)
         if not instances:
             module.fail_json(msg="There are no instances in our record based on instance_ids {0}. "
                                  "Please check it and try again.".format(instance_ids))